eapis: Reorganize files for new build system
- Create eapis/include/hve and move hve headers there
- Update include paths to reflect new structure
- Use module-qualified include paths to disambiguate eapi includes from
  hypervisor includes (i.e. include <bfvmm/include/...>).
