+++ include/hve/arch/intel_x64/vmcs/ept.h
@@ -30,7 +30,7 @@
 #include <memory>
 
 #include <bfmemory.h>
-#include <vmcs/ept_entry_intel_x64.h>
+#include <hve/arch/intel_x64/vmcs/ept_entry.h>
 
 // -----------------------------------------------------------------------------
 // Exports
