+++ include/hve/arch/intel_x64/vmcs/vmcs_vmm_state.h
@@ -1,5 +1,5 @@
 //
-// Bareflank Extended APIs
+// Bareflank Hypervisor
 //
 // Copyright (C) 2015 Assured Information Security, Inc.
 // Author: Rian Quinn        <quinnr@ainfosec.com>
@@ -19,22 +19,11 @@
 // License along with this library; if not, write to the Free Software
 // Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA
 
-#ifndef EXIT_HANDLER_INTEL_X64_EAPIS_MSR_VERIFIERS_H
-#define EXIT_HANDLER_INTEL_X64_EAPIS_MSR_VERIFIERS_H
+#ifndef VMCS_INTEL_X64_VMM_STATE_EAPIS_H
+#define VMCS_INTEL_X64_VMM_STATE_EAPIS_H
 
-#include <exit_handler/exit_handler_intel_x64_eapis.h>
-#include <exit_handler/exit_handler_intel_x64_eapis_verifiers.h>
-
-/// @cond
-
-// -----------------------------------------------------------------------------
-// Constants
-// -----------------------------------------------------------------------------
-
-namespace vp
-{
-constexpr const auto index_enable_msr_bitmap = 0x0003001UL;
-}
+#include <bfvmm/include/hve/arch/x64/idt.h>
+#include <bfvmm/include/hve/arch/intel_x64/vmcs/vmcs_state_vmm.h>
 
 // -----------------------------------------------------------------------------
 // Exports
@@ -42,31 +31,50 @@ constexpr const auto index_enable_msr_bitmap = 0x0003001UL;
 
 #include <bfexports.h>
 
-#ifndef STATIC_EAPIS_EXIT_HANDLER
-#ifdef SHARED_EAPIS_EXIT_HANDLER
-#define EXPORT_EAPIS_EXIT_HANDLER EXPORT_SYM
+#ifndef STATIC_EAPIS_VMCS
+#ifdef SHARED_EAPIS_VMCS
+#define EXPORT_EAPIS_VMCS EXPORT_SYM
 #else
-#define EXPORT_EAPIS_EXIT_HANDLER IMPORT_SYM
+#define EXPORT_EAPIS_VMCS IMPORT_SYM
 #endif
 #else
-#define EXPORT_EAPIS_EXIT_HANDLER
+#define EXPORT_EAPIS_VMCS
 #endif
 
 // -----------------------------------------------------------------------------
 // Definitions
 // -----------------------------------------------------------------------------
 
-class EXPORT_EAPIS_EXIT_HANDLER default_verifier__enable_msr_bitmap :
-    public vmcall_verifier
+/// VMM Intel x64 State (EAPIs)
+///
+/// The following defines the state class for the VMM. The default VMM state
+/// is used here except the constructor, were EAPI specific resources are
+/// initialized.
+///
+class EXPORT_EAPIS_VMCS vmcs_intel_x64_vmm_state_eapis :
+    public vmcs_intel_x64_vmm_state
 {
 public:
-    default_verifier__enable_msr_bitmap() = default;
-    ~default_verifier__enable_msr_bitmap() override = default;
 
-    virtual verifier_result verify(bool enabled)
-    { bfignored(enabled); return default_verify(); }
+    /// Default Constructor
+    ///
+    vmcs_intel_x64_vmm_state_eapis();
+
+    /// Default Destructor
+    ///
+    ~vmcs_intel_x64_vmm_state_eapis() override = default;
+
+public:
+
+    /// @cond
+
+    vmcs_intel_x64_vmm_state_eapis(vmcs_intel_x64_vmm_state_eapis &&) noexcept = delete;
+    vmcs_intel_x64_vmm_state_eapis &operator=(vmcs_intel_x64_vmm_state_eapis &&) noexcept = delete;
+
+    vmcs_intel_x64_vmm_state_eapis(const vmcs_intel_x64_vmm_state_eapis &) = delete;
+    vmcs_intel_x64_vmm_state_eapis &operator=(const vmcs_intel_x64_vmm_state_eapis &) = delete;
+
+    /// @endcond
 };
-
-/// @endcond
 
 #endif
