+++ src/hve/arch/intel_x64/exit_handler/vpid_vmcall.cpp
@@ -19,11 +19,11 @@
 // License along with this library; if not, write to the Free Software
 // Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA
 
-#include <exit_handler/exit_handler_intel_x64_eapis.h>
-#include <exit_handler/exit_handler_intel_x64_eapis_vmcall_interface.h>
+#include <hve/arch/intel_x64/exit_handler/exit_handler.h>
+#include <hve/arch/intel_x64/exit_handler/vmcall_interface.h>
 
-#include <exit_handler/exit_handler_intel_x64_eapis_verifiers.h>
-#include <exit_handler/exit_handler_intel_x64_eapis_vpid_verifiers.h>
+#include <hve/arch/intel_x64/exit_handler/verifiers.h>
+#include <hve/arch/intel_x64/exit_handler/vpid_verifiers.h>
 
 using namespace x64;
 using namespace intel_x64;
