+++ src/hve/arch/intel_x64/vmcs/apic.cpp
@@ -20,10 +20,8 @@
 // Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA
 
 #include <bfdebug.h>
-#include <vmcs/apic_intel_x64.h>
-
-#include <arch/intel_x64/cpuid.h>
-#include <arch/x64/misc.h>
+#include <bfintrinsics/include/arch/intel_x64/cpuid.h>
+#include <hve/arch/intel_x64/vmcs/apic.h>
 
 using namespace x64;
 using namespace intel_x64;
