+++ src/hve/arch/intel_x64/vmcs/vmcs.cpp
@@ -19,7 +19,7 @@
 // License along with this library; if not, write to the Free Software
 // Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA
 
-#include <vmcs/vmcs_intel_x64_eapis.h>
+#include <hve/arch/intel_x64/vmcs/vmcs.h>
 
 vmcs_intel_x64_eapis::vmcs_intel_x64_eapis()
 {
