+++ src/hve/arch/intel_x64/vmcs/vmcs_cr_access_hooks.cpp
@@ -19,10 +19,10 @@
 // License along with this library; if not, write to the Free Software
 // Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA
 
-#include <vmcs/vmcs_intel_x64_eapis.h>
-#include <arch/intel_x64/vmcs/16bit_control_fields.h>
-#include <arch/intel_x64/vmcs/32bit_control_fields.h>
-#include <arch/intel_x64/vmcs/natural_width_control_fields.h>
+#include <hve/arch/intel_x64/vmcs/vmcs.h>
+#include <bfintrinsics/include/arch/intel_x64/vmcs/16bit_control_fields.h>
+#include <bfintrinsics/include/arch/intel_x64/vmcs/32bit_control_fields.h>
+#include <bfintrinsics/include/arch/intel_x64/vmcs/natural_width_control_fields.h>
 
 using namespace intel_x64;
 using namespace vmcs;
